+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Fix removing Jellyfin and Emby libraries that have been deleted from the source media server
+
 ### Added
 - Add audio stream selector scripts for episodes and movies
   - This will let you customize which audio stream is selected for playback